+++ src/ralph/api/auth/user.py
@@ -1,6 +1,6 @@
 """Authenticated user for the Ralph API."""
 
-from typing import List, Optional
+from typing import Dict, List
 
 from pydantic import BaseModel
 
@@ -9,9 +9,9 @@ class AuthenticatedUser(BaseModel):
     """Pydantic model for user authentication.
 
     Attributes:
-        username (str): Consists of the username of the current user.
-        scopes (list): Consists of the scopes the user has access to.
+        agent (dict): The agent representing the current user.
+        scopes (list): The scopes the user has access to.
     """
 
-    username: str
-    scopes: Optional[List[str]]
+    agent: Dict
+    scopes: List[str]